refactor: move channels_v2/sureadhere_channel.py to channels/sureadhere_channel.py
